+++ app/src/main/java/moneytraker/com/moneytracker/TempActivity.java
@@ -12,7 +12,7 @@ public class TempActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.add_activity);
 
         final TextView add = (TextView) findViewById(R.id.add);
