+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,14 @@
     <string name="price_hint">0</string>
     <string name="plus">+</string>
     <string name="rouble">\u20bd</string>
+
+    <string name="expenses">Расходы</string>
+    <string name="incomes">Доходы</string>
+    <string name="balance">Баланс</string>
+
+    <array name="main_pager_titles">
+        <item>@string/expenses</item>
+        <item>@string/incomes</item>
+        <item>@string/balance</item>
+    </array>
 </resources>
